Add Verification Chat Modification | Shuffle Some Code Around |
- Move some bot code around in order to accommodate some logic regarding bot querying. (I made no solid way to pass the bot scope to the express server)
- Add bot querying logic. Essentially, add requests to where the server can query bot specific values. Case for adding this was getting the discriminator of the user.
- Add more request handlers in the plugin in order to accommodate these new requests.
- Add configuration value: verified-prefix. It can take either a true, or false, and determines whether or not the server will modify chat messages to indicate the player's verification status.
- Add event to check if a player's verified, and if so modify the chat to indicate if they are. It checks the player's preferences, which segues into the next intended update.

For the next update, I want to add a player preference GUI.
Design:
/prefs || /preferences
Opens a gui, with 3 blocks. If the preference for the player is on, it's a green clay, for off it's a red clay. They click it to modify the value in the gui. There's then a save button in the bottom right, and a close button in the bottom left.

My current plans right now is to only save the preferences if the player clicks the save button, but I may change this in later updates.

NOTE:
The major feature added in this update NEEDS to be stress testing. I saw a noticeable .1ms of delay between chat being sent, and viewed. I don't think it will be an issue as my development environment was stressed to the max with a lot of other things, but I just want to ensure the feature is optimized.